requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==5.0.7
filelock==3.15.4
platformdirs==4.2.2
sqlparse==0.5.1
tzdata==2024.1
virtualenv==20.26.3


certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
defusedxml==0.8.0rc2
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.7
mysqlclient==2.2.4
oauthlib==3.2.2
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4

urllib3==2.2.2
